@@ -1,13 +1,13 @@
 apply plugin: 'com.android.application'
 
 android {
-	compileSdkVersion 27
+	compileSdkVersion 28
 	defaultConfig {
 		applicationId "com.nfcalarmclock"
 		minSdkVersion 21
-		targetSdkVersion 27
-		versionCode 78
-		versionName "8.2.10"
+		targetSdkVersion 28
+		versionCode 79
+		versionName "8.2.11"
 		testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
 	}
 	buildTypes {
@@ -36,11 +36,22 @@ repositories {
 
 dependencies {
 	implementation fileTree(dir: 'libs', include: ['*.jar'])
-	implementation 'com.android.support:appcompat-v7:27.1.1'
-	implementation 'com.android.support:support-v4:27.1.1'
-	implementation 'com.android.support:design:27.1.1'
-	implementation 'com.android.support:recyclerview-v7:27.1.1'
-	implementation 'com.android.support:cardview-v7:27.1.1'
+	implementation 'androidx.annotation:annotation:1.1.0'
+	implementation 'androidx.appcompat:appcompat:1.1.0'
+	implementation 'androidx.cardview:cardview:1.0.0'
+	implementation 'androidx.coordinatorlayout:coordinatorlayout:1.0.0'
+	implementation 'androidx.core:core:1.1.0'
+	implementation 'androidx.fragment:fragment:1.1.0'
+	implementation 'androidx.preference:preference:1.1.0'
+	implementation 'androidx.recyclerview:recyclerview:1.0.0'
+	implementation 'androidx.viewpager:viewpager:1.0.0'
+	implementation 'com.google.android.material:material:1.0.0'
+
+	//implementation 'com.android.support:appcompat-v7:28.0.0'
+	//implementation 'com.android.support:support-v4:28.0.0'
+	//implementation 'com.android.support:design:28.0.0'
+	//implementation 'com.android.support:recyclerview-v7:28.0.0'
+	//implementation 'com.android.support:cardview-v7:28.0.0'
 
 	implementation 'com.spotify.sdk:spotify-auth-release:1.1.0@aar'
 	implementation 'com.spotify.sdk:spotify-app-remote-release:0.6.1@aar'
