@@ -1,16 +1,21 @@
 package com.nfcalarmclock;
 
-import android.app.FragmentManager;
+//import android.app.FragmentManager;
+//import android.support.v4.app.FragmentManager;
 import android.content.Context;
 import android.content.SharedPreferences;
-import android.preference.Preference;
-import android.preference.PreferenceFragment;
+//import android.preference.Preference;
+//import android.preference.PreferenceFragment;
+
+import androidx.fragment.app.FragmentManager;
+import androidx.preference.Preference;
+import androidx.preference.PreferenceFragmentCompat;
 
 /**
  * Settings fragment.
  */
 public abstract class NacSettingsFragment
-	extends PreferenceFragment
+	extends PreferenceFragmentCompat
 	implements SharedPreferences.OnSharedPreferenceChangeListener
 {
 
